A Red Themed Website
Improvements
Build a red-themed new Detectovid site.Retained the original organization	of the site content.Texts (except headings) are aligned as “justify”.Fixed a few places where quotation marks are missing in the original code.Fixed network icon colors on the team page.Organized images (in source code) based on team, edu and organization.Use the new project logo.Incorporated the latest donation links.

Items may need adjustment
The banner picture is a placeholder: We detect that the rain is coming and will be well prepared when it arrives. Further out, the Sun is rising. A bright day is ahead of us.In the new implementation, we use CSS variables in a few place. All browsers around us support CSS variables. If that turns out to be a problem, the variables can be easily string-replaced with their actual values.